코드 정리

+++ src/main/java/com/example/demo/domain/card/dto/request/CardRequest.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class CardRequest {
 
-  Integer score;
+  Long score;
   String imageUrl;
   String title;
   String subtitle;
